Include exit-1 commamds in history
The success/error status of the command should not determine whether the
command is added to the history. Some commands could exit 1 sometimes,
but other times, they could exit 0 (e.g. running unit tests). However,
it is still desirable that such commands be added to the history so they
can be easily re-run.